+++ src/Test/Tensor.hs
@@ -29,12 +29,13 @@ module Test.Tensor (
   , sizeAtLeast
     -- * Standard operations
   , zipWith
+  , zipEach
   , replicate
   , rotate
   , distrib
   , transpose
-  , foreach
-  , foreachWith
+  , forEach
+  , forEachWith
     -- * Subtensors
   , subs
   , subsWithStride
@@ -101,7 +102,7 @@ import Test.QuickCheck qualified as QC
 --
 -- Invariants:
 --
--- * The dimension must be strictly positive (zero is not allowed)
+-- * None of the dimensions may be empty
 -- * Tensors must be rectangular
 --
 -- (These invariants could in principle be enforced by using more precise types,
@@ -129,7 +130,7 @@ getTensor (Tensor xs) = xs
 type Size n = Vec n Int
 
 -- | Analogue of 'List.length'
-size :: Tensor n a -> Size n
+size :: HasCallStack => Tensor n a -> Size n
 size (Scalar _)  = VNil
 size (Tensor xs) = L.length xs ::: size (L.head xs)
 
@@ -145,6 +146,15 @@ sizeAtLeast sz = and . Foldable.toList . Vec.zipWith (<=) sz . size
 zipWith :: (a -> b -> c) -> Tensor n a -> Tensor n b -> Tensor n c
 zipWith f (Scalar a)  (Scalar b)  = Scalar (f a b)
 zipWith f (Tensor as) (Tensor bs) = Tensor $ L.zipWith (zipWith f) as bs
+
+-- | Zip each top-level dimension
+--
+-- Unlike 'zipWith', this allows for two tensors of different dimensions,
+-- somewhat similar to 'forEach'.
+zipEach ::
+     (Tensor n a -> Tensor m b -> Tensor l c)
+  -> Tensor (S n) a -> Tensor (S m) b -> Tensor (S l) c
+zipEach f (Tensor as) (Tensor bs) = Tensor $ L.zipWith f as bs
 
 -- | Analogue of 'List.replicate'
 replicate :: Size n -> a -> Tensor n a
@@ -173,16 +183,16 @@ transpose :: Tensor Nat2 a -> Tensor Nat2 a
 transpose = fromLists . L.transpose . toLists
 
 -- | Map element over the first dimension of the tensor
-foreach :: Tensor (S n) a -> (Tensor n a -> Tensor m b) -> Tensor (S m) b
-foreach (Tensor as) f = Tensor (Prelude.map f as)
+forEach :: Tensor (S n) a -> (Tensor n a -> Tensor m b) -> Tensor (S m) b
+forEach (Tensor as) f = Tensor (Prelude.map f as)
 
--- | Variation of 'foreach' with an auxiliary list
-foreachWith ::
+-- | Variation of 'forEach' with an auxiliary list
+forEachWith ::
     Tensor (S n) a
  -> [x]
  -> (Tensor n a -> x -> Tensor m b)
  -> Tensor (S m) b
-foreachWith (Tensor as) xs f = Tensor (L.zipWith f as xs)
+forEachWith (Tensor as) xs f = Tensor (L.zipWith f as xs)
 
 {-------------------------------------------------------------------------------
   Subtensors
@@ -584,22 +594,39 @@ showConstructor p sn
     n' = snatToNatural sn
 
 instance Show a => Show (Tensor n a) where
-  showsPrec p tensor = showLists (Proxy @a) (tensorSNat tensor) $
-      showParen (p >= appPrec1) $
-          showConstructor appPrec1 (tensorSNat tensor)
-        . showSpace
-        . showsPrec appPrec1 (toLists tensor)
+  showsPrec p tensor = showParen (p >= appPrec1) $
+      case tensorSNat tensor of
+        Just sz ->
+          showLists (Proxy @a) sz $
+              showConstructor appPrec1 sz
+            . showSpace
+            . showsPrec appPrec1 (toLists tensor)
+        Nothing ->
+            showString "fromLists "
+          . showInvalid tensor
+    where
+      -- Tensors with empty dimensions are invalid, but if we fail to 'show'
+      -- them, that would make debugging code that accidentally creates them
+      -- unnecessarily difficult. We therefore special case this.
+      showInvalid :: forall m. Tensor m a -> ShowS
+      showInvalid (Scalar x)  = showsPrec 0 x
+      showInvalid (Tensor xs) =
+            showString "["
+          . L.foldr (.) id (L.intersperse (showString ",") (map showInvalid xs))
+          . showString "]"
 
 {-------------------------------------------------------------------------------
   Internal auxiliary: SNat
 -------------------------------------------------------------------------------}
 
-tensorSNatI :: Tensor n a -> (SNatI n => r) -> r
-tensorSNatI (Scalar _)  k = k
-tensorSNatI (Tensor xs) k = tensorSNatI (L.head xs) k
+-- | Size of the tensor, unless it contains empty dimensions
+tensorSNatI :: Tensor n a -> r -> (SNatI n => r) -> r
+tensorSNatI (Scalar _    ) _    kOk = kOk
+tensorSNatI (Tensor (x:_)) kErr kOk = tensorSNatI x kErr kOk
+tensorSNatI (Tensor []   ) kErr _   = kErr
 
-tensorSNat :: Tensor n a -> SNat n
-tensorSNat tensor = tensorSNatI tensor snat
+tensorSNat :: Tensor n a -> Maybe (SNat n)
+tensorSNat tensor = tensorSNatI tensor Nothing (Just snat)
 
 {-------------------------------------------------------------------------------
   Internal auxiliary: lists
@@ -647,6 +674,6 @@ pickOne = \case
     go acc x (y:ys) = (reverse acc, x, (y:ys)) : go (x:acc) y ys
 
 -- | Distribute @f@ over @[]@
-distribList :: Functor f => Int -> f [a] -> [f a]
+distribList :: (HasCallStack, Functor f) => Int -> f [a] -> [f a]
 distribList 0 _   = []
-distribList n fxs = (head <$> fxs) : distribList (n - 1) (tail <$> fxs)
+distribList n fxs = (L.head <$> fxs) : distribList (n - 1) (tail <$> fxs)